Add GET handler for episodios endpoint
